criacao do endpoint de editar pedido e endpoint lista pedidos por mesa

+++ backend/src/main/java/com/tarcio4lmeida/barclick/entities/ItemPedido.java
@@ -2,6 +2,8 @@ package com.tarcio4lmeida.barclick.entities;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "tb_item_pedido")
 public class ItemPedido {
@@ -9,7 +11,7 @@ public class ItemPedido {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private Integer quantidade;
-    private Double preco;
+    private BigDecimal preco;
 
     @ManyToOne
     @JoinColumn(name = "produto_id")
@@ -35,11 +37,11 @@ public class ItemPedido {
         this.quantidade = quantidade;
     }
 
-    public Double getPreco() {
+    public BigDecimal getPreco() {
         return preco;
     }
 
-    public void setPreco(Double preco) {
+    public void setPreco(BigDecimal preco) {
         this.preco = preco;
     }
